Move libcore testing classes into testing
When conscrypt was first separated from AOSP some of the AOSP libcore
classes were copied into libcore-stub which is compiled against for
both the openjdk and platform builds but only used at runtime by the
openjdk builds. Unfortunately, those classes clash with internal classes
provided by the Android runtime which will cause problems in future
releases.

This change moves those libcore-stub classes that are only used for
testing from their libcore.... package into a org.conscrypt.... package
to prevent clashes. It also moves them out of the libcore-stub module
and into the testing module so they will be used on both platform and
openjdk builds.
