+++ openjdk-integ-tests/src/test/java/org/conscrypt/java/security/SignatureTest.java
@@ -66,7 +66,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.concurrent.Future;
 import java.util.concurrent.TimeUnit;
-import libcore.java.security.StandardNames;
 import org.conscrypt.Conscrypt;
 import org.conscrypt.TestUtils;
 import org.junit.Test;
