+++ openjdk-integ-tests/src/test/java/org/conscrypt/javax/net/ssl/SSLServerSocketTest.java
@@ -25,8 +25,8 @@ import java.util.HashSet;
 import javax.net.ssl.SSLContext;
 import javax.net.ssl.SSLServerSocket;
 import javax.net.ssl.SSLServerSocketFactory;
-import libcore.java.security.StandardNames;
 import org.conscrypt.TestUtils;
+import org.conscrypt.java.security.StandardNames;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.JUnit4;
