+++ openjdk-integ-tests/src/test/java/org/conscrypt/javax/net/ssl/SSLSessionTest.java
@@ -39,8 +39,8 @@ import javax.net.ssl.SSLSession;
 import javax.net.ssl.SSLSessionBindingEvent;
 import javax.net.ssl.SSLSessionBindingListener;
 import javax.net.ssl.SSLSocket;
-import libcore.java.security.StandardNames;
 import org.conscrypt.TestUtils;
+import org.conscrypt.java.security.StandardNames;
 import org.conscrypt.java.security.TestKeyStore;
 import org.junit.Test;
 import org.junit.runner.RunWith;
