+++ openjdk/src/test/java/org/conscrypt/NativeCryptoTest.java
@@ -78,9 +78,9 @@ import javax.net.ssl.SSLHandshakeException;
 import javax.net.ssl.SSLProtocolException;
 import javax.security.auth.x500.X500Principal;
 import libcore.io.IoUtils;
-import libcore.java.security.StandardNames;
 import org.conscrypt.NativeCrypto.SSLHandshakeCallbacks;
 import org.conscrypt.OpenSSLX509CertificateFactory.ParsingException;
+import org.conscrypt.java.security.StandardNames;
 import org.conscrypt.java.security.TestKeyStore;
 import org.junit.BeforeClass;
 import org.junit.Test;
