+++ testing/src/main/java/org/conscrypt/java/security/StandardNames.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package libcore.java.security;
+package org.conscrypt.java.security;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
