+++ testing/src/main/java/org/conscrypt/testing/Streams.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package libcore.io;
+package org.conscrypt.testing;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
